Move check for running Docker from DotNet to SamDebug base class
